requirements.txt
google-adk
mcp
python-dotenv 
mysql-connector-python
google-generativeai